requirements.txt
# required to run
pillow==10.1.0
exifread==3.0.0
pydantic==2.5.3
click==7.1.2

# test
pytest==7.4.0
pytest-cov==4.1.0
pytest-mock==3.12.0
piexif==1.1.3

# formatting
pre-commit==3.5.0

# building package
build==1.0.3